fix response and bug no photo was selected but error

+++ features/attendance/attendanceEntity.go
@@ -3,17 +3,17 @@ package attendance
 import "github.com/labstack/echo/v4"
 
 type Core struct {
-	ID               uint   `json:"id"`
-	AttendanceDate   string `json:"attendance_date"`
-	ClockIn          string `json:"clock_in"`
-	ClockInLocation  string `json:"clock_in_location"`
-	ClockInOSM       string `json:"clock_in_map_location"`
-	ClockOut         string `json:"clock_out"`
-	ClockOutLocation string `json:"clock_out_location"`
-	ClockOutOSM      string `json:"clock_out_map_location"`
-	Attendance       string `json:"attendance"`
-	AttendanceStatus string `json:"attendance_status"`
-	WorkTime         int    `json:"work_time"`
+	ID               uint    `json:"id"`
+	AttendanceDate   string  `json:"attendance_date"`
+	ClockIn          string  `json:"clock_in"`
+	ClockInLocation  string  `json:"clock_in_location"`
+	ClockInOSM       string  `json:"clock_in_map_location"`
+	ClockOut         string  `json:"clock_out"`
+	ClockOutLocation string  `json:"clock_out_location"`
+	ClockOutOSM      string  `json:"clock_out_map_location"`
+	Attendance       string  `json:"attendance"`
+	AttendanceStatus string  `json:"attendance_status"`
+	WorkTime         float32 `json:"work_time"`
 }
 
 type AttendanceHandler interface {
@@ -26,7 +26,6 @@ type AttendanceHandler interface {
 	GetPresenceTotalToday() echo.HandlerFunc
 	GetPresenceDetail() echo.HandlerFunc
 	RecordByID() echo.HandlerFunc
-	// Graph() echo.HandlerFunc
 }
 
 type AttendanceService interface {
@@ -38,7 +37,6 @@ type AttendanceService interface {
 	GetPresenceTotalToday(token interface{}) ([]Core, error)
 	GetPresenceDetail(token interface{}, attendanceID uint) (Core, error)
 	RecordByID(employeeID uint, dateFrom string, dateTo string) ([]Core, string, error)
-	// Graph(param string) ([]Core, error)
 }
 
 type AttendanceData interface {
@@ -49,5 +47,4 @@ type AttendanceData interface {
 	GetPresenceToday(employeeID uint) (Core, error)
 	GetPresenceTotalToday(adminID uint) ([]Core, error)
 	GetPresenceDetail(adminID uint, attendanceID uint) (Core, error)
-	// Graph(param string) ([]Core, error)
 }
